LibGfx: Rename QMArithmeticDecoder.{h,cpp} to QMArithmeticCoder.{h,cpp}
In preparation for adding a QMArithmeticEncoder to it.

No behavior change.
