+++ Userland/Libraries/LibGfx/CMakeLists.txt
@@ -71,7 +71,7 @@ set(SOURCES
     ImageFormats/PortableFormatWriter.cpp
     ImageFormats/PAMLoader.cpp
     ImageFormats/PPMLoader.cpp
-    ImageFormats/QMArithmeticDecoder.cpp
+    ImageFormats/QMArithmeticCoder.cpp
     ImageFormats/QOILoader.cpp
     ImageFormats/QOIWriter.cpp
     ImageFormats/TGALoader.cpp
