+++ Userland/Libraries/LibGfx/ImageFormats/QMArithmeticCoder.h
